learningmodule/requirements.txt
fastapi>=0.115
uvicorn>=0.29
pronouncing>=0.2.0
pyttsx3>=2.90
python-dotenv>=1.0
pydub>=0.25.1